Course dashboard (#86)
* Initial version of the course selection page, can be reached on /courses. Uses mockdata

* API debugging

* Export interfaces in types.d.ts

* Revert "Export interface fix (#47)"

This reverts commit 646ad4487cd327cb8d9ef42588fb78147ac1284c.

* Revert "Revert "Export interface fix (#47)""

This reverts commit 76b84627af842975d1ff0e4c498ecfd0c6b95d2f.

* Working and semi-styled course page, also lint proof.

Also already contains the possibility to separate between student courses or TA courses but that is currently unused

* Added the missing CourseList component

* Card layout for courses

* Small fix + lint

* Removed CourseSelectAction, from now on it just pushes the URL directly.

* Suggested change to the course state: leave out course

This is probably a better state model. Since CourseDtoFull will be an extension on CourseDtoSummary we can just use Type Guards to check whether we already have the full version or only the summary and then request the full version if it is not yet there. This also increases effiency when regularly switching between courses

* Some selectors, actions and constants for the course dashboard

* Course dashboard main content

* CourseDashboard responsive sidebar and improved responsiveness for cards

* Added sidebar icon placeholder and cleaned code